@@ -1,6 +1,6 @@
 module github.com/mattermost-community/mattermost-plugin-google-drive
 
-go 1.21
+go 1.24
 
 require (
 	github.com/golang/mock v1.6.0
