@@ -16,7 +16,7 @@ import (
 )
 
 type Tracker interface {
-	TrackUserEvent(event, userID string, properties map[string]interface{})
+	TrackUserEvent(event, userID string, properties map[string]any)
 }
 
 type FlowManager struct {
@@ -147,14 +147,14 @@ func (fm *FlowManager) StartSetupWizard(userID string) error {
 }
 
 func (fm *FlowManager) trackStartSetupWizard(userID string, fromInvite bool) {
-	fm.tracker.TrackUserEvent("setup_wizard_start", userID, map[string]interface{}{
+	fm.tracker.TrackUserEvent("setup_wizard_start", userID, map[string]any{
 		"from_invite": fromInvite,
 		"time":        model.GetMillis(),
 	})
 }
 
 func (fm *FlowManager) trackCompleteSetupWizard(userID string) {
-	fm.tracker.TrackUserEvent("setup_wizard_complete", userID, map[string]interface{}{
+	fm.tracker.TrackUserEvent("setup_wizard_complete", userID, map[string]any{
 		"time": model.GetMillis(),
 	})
 }
@@ -173,13 +173,13 @@ func (fm *FlowManager) StartOauthWizard(userID string) error {
 }
 
 func (fm *FlowManager) trackStartOauthWizard(userID string) {
-	fm.tracker.TrackUserEvent("oauth_wizard_start", userID, map[string]interface{}{
+	fm.tracker.TrackUserEvent("oauth_wizard_start", userID, map[string]any{
 		"time": model.GetMillis(),
 	})
 }
 
 func (fm *FlowManager) trackCompleteOauthWizard(userID string) {
-	fm.tracker.TrackUserEvent("oauth_wizard_complete", userID, map[string]interface{}{
+	fm.tracker.TrackUserEvent("oauth_wizard_complete", userID, map[string]any{
 		"time": model.GetMillis(),
 	})
 }
@@ -204,7 +204,8 @@ func (fm *FlowManager) stepOAuthInfo() flow.Step {
 ##### :white_check_mark: Step 1: Register an OAuth Application in Google Cloud Console
 You must first register the Mattermost Google Drive Plugin as an authorized OAuth app.`
 	oauthMessage := fmt.Sprintf(
-		"1. Create a new Project. You would need to redirect to [Google Cloud Console](https://console.cloud.google.com/home/dashboard) and select the option to **New project**. Then, select the name and the organization (optional).\n"+
+		"1. Create a new Project. You would need to redirect to [Google Cloud Console](https://console.cloud.google.com/home/dashboard) and select the option to **New project**. Then, select the name and the organization (optional).\n"+ //nolint:unqueryvet
+
 			"2. Select APIs. After creating a project, on the left side menu on **APIs & Services**, then, select the first option **Enabled APIs & Services** and wait, the page will redirect.\n"+
 			"3. Click on **Enable APIs and Services** option. Once you are in the API Library, search and enable following APIs:\n"+
 			"	- Google Drive API\n"+
@@ -261,7 +262,7 @@ func (fm *FlowManager) stepOAuthInput() flow.Step {
 		WithButton(cancelButton())
 }
 
-func (fm *FlowManager) submitOAuthConfig(f *flow.Flow, submitted map[string]interface{}) (flow.Name, flow.State, map[string]string, error) {
+func (fm *FlowManager) submitOAuthConfig(f *flow.Flow, submitted map[string]any) (flow.Name, flow.State, map[string]string, error) {
 	errorList := map[string]string{}
 
 	clientIDRaw, ok := submitted["client_id"]
@@ -333,7 +334,7 @@ func (fm *FlowManager) StartAnnouncementWizard(userID string) error {
 }
 
 func (fm *FlowManager) trackStartAnnouncementWizard(userID string) {
-	fm.tracker.TrackUserEvent("announcement_wizard_start", userID, map[string]interface{}{
+	fm.tracker.TrackUserEvent("announcement_wizard_start", userID, map[string]any{
 		"time": model.GetMillis(),
 	})
 }
@@ -384,7 +385,7 @@ func (fm *FlowManager) stepAnnouncementConfirmation() flow.Step {
 		OnRender(func(f *flow.Flow) { fm.trackCompleteAnnouncementWizard(f.UserID) })
 }
 
-func (fm *FlowManager) submitChannelAnnouncement(f *flow.Flow, submitted map[string]interface{}) (flow.Name, flow.State, map[string]string, error) {
+func (fm *FlowManager) submitChannelAnnouncement(f *flow.Flow, submitted map[string]any) (flow.Name, flow.State, map[string]string, error) {
 	channelIDRaw, ok := submitted["channel_id"]
 	if !ok {
 		return "", nil, nil, errors.New("channel_id missing")
@@ -424,7 +425,7 @@ func (fm *FlowManager) submitChannelAnnouncement(f *flow.Flow, submitted map[str
 }
 
 func (fm *FlowManager) trackCompleteAnnouncementWizard(userID string) {
-	fm.tracker.TrackUserEvent("announcement_wizard_complete", userID, map[string]interface{}{
+	fm.tracker.TrackUserEvent("announcement_wizard_complete", userID, map[string]any{
 		"time": model.GetMillis(),
 	})
 }
