@@ -96,8 +96,8 @@ func TestNewCommentReply(t *testing.T) {
 		te.mockAPI.On("GetConfig").Return(&mattermostModel.Config{ServiceSettings: mattermostModel.ServiceSettings{SiteURL: &siteURL}})
 		te.mockAPI.On("GetDirectChannel", "userID1", te.plugin.BotUserID).Return(&mattermostModel.Channel{Id: "channelId1"}, nil).Times(1)
 		post := GetDMPost(te.plugin.BotUserID, comment, file, siteURL)
-		post.Props["attachments"].([]interface{})[0].(map[string]interface{})["pretext"] = fmt.Sprintf("User 1 replied on %s %s", utils.GetInlineImage("File icon:", file.IconLink), utils.GetHyperlink(file.Name, ""))
-		post.Props["attachments"].([]interface{})[0].(map[string]interface{})["text"] = fmt.Sprintf("Previous reply:\n%s\n> %s", comment.Content, "Reply to comment")
+		post.Props["attachments"].([]any)[0].(map[string]any)["pretext"] = fmt.Sprintf("User 1 replied on %s %s", utils.GetInlineImage("File icon:", file.IconLink), utils.GetHyperlink(file.Name, ""))
+		post.Props["attachments"].([]any)[0].(map[string]any)["text"] = fmt.Sprintf("Previous reply:\n%s\n> %s", comment.Content, "Reply to comment")
 		te.mockAPI.On("CreatePost", post).Return(nil, nil).Times(1)
 		te.plugin.handleReplyAdded(context.Background(), mockGoogleDrive, fileID, "userID1", activity.Activities[0], file)
 	})
@@ -109,8 +109,8 @@ func TestNewCommentReply(t *testing.T) {
 		te.mockAPI.On("GetConfig").Return(&mattermostModel.Config{ServiceSettings: mattermostModel.ServiceSettings{SiteURL: &siteURL}})
 		te.mockAPI.On("GetDirectChannel", "userID1", te.plugin.BotUserID).Return(&mattermostModel.Channel{Id: "channelId1"}, nil).Times(1)
 		post := GetDMPost(te.plugin.BotUserID, comment, file, siteURL)
-		post.Props["attachments"].([]interface{})[0].(map[string]interface{})["pretext"] = fmt.Sprintf("User 1 replied on %s %s", utils.GetInlineImage("File icon:", file.IconLink), utils.GetHyperlink(file.Name, ""))
-		post.Props["attachments"].([]interface{})[0].(map[string]interface{})["text"] = fmt.Sprintf("Previous reply:\n%s\n> %s", comment.Content, utils.MarkdownToHTMLEntities("> Reply to comment"))
+		post.Props["attachments"].([]any)[0].(map[string]any)["pretext"] = fmt.Sprintf("User 1 replied on %s %s", utils.GetInlineImage("File icon:", file.IconLink), utils.GetHyperlink(file.Name, ""))
+		post.Props["attachments"].([]any)[0].(map[string]any)["text"] = fmt.Sprintf("Previous reply:\n%s\n> %s", comment.Content, utils.MarkdownToHTMLEntities("> Reply to comment"))
 		te.mockAPI.On("CreatePost", post).Return(nil, nil).Times(1)
 		te.plugin.handleReplyAdded(context.Background(), mockGoogleDrive, fileID, "userID1", activity.Activities[0], file)
 	})
